Move template tests
